+++ donseats-frontend/src/App.js
@@ -6,13 +6,16 @@ import Signup from './components/Signup';
 import Menu from './components/Menu';
 import BonsMenu from './components/BonsMenu';
 import DonsMenu from './components/DonsMenu';
+import JavaMenu from './components/JavaMenu';
 import BonsNavBar from './components/BonsNavBar';
 import DonsNavBar from './components/DonsNavBar';
+import JavaNavBar from './components/JavaNavBar';
 import { AuthProvider } from './services/AuthContext'; 
 import { AuthContext } from './services/AuthContext'; 
 import Dashboard from './components/Dashboard';
 import BonsDashboard from './components/BonsDashboard';
 import DonsDashboard from './components/DonsDashboard';
+import JavaDashboard from './components/JavaDashboard';
 import { useContext } from 'react'; 
 import axios from 'axios'; 
 import NavBar from './components/NavBar';
@@ -28,10 +31,12 @@ const App = () => {
   // const [selectedCategory, setSelectedCategory] = useState('Egg Sandwiches');
   const [activeCategory, setActiveCategory] = useState('Egg Sandwiches');
   const [bonsSelectedCategory, setBonsSelectedCategory] = useState('Coffee & Espresso');
-  const [donsSelectedCategory, setDonsSelectedCategory] = useState('Epic Eats'); // Default for Bon Bon's
+  const [donsSelectedCategory, setDonsSelectedCategory] = useState('Epic Eats');
+  const [javaSelectedCategory, setJavaSelectedCategory] = useState('Espresso'); // Default for Bon Bon's
   const [einsteinCartItems, setEinsteinCartItems] = useState({});  // Einstein's cart
   const [bonsCartItems, setBonsCartItems] = useState({});  
-  const [donsCartItems, setDonsCartItems] = useState({});        // Bon Bon's cart
+  const [donsCartItems, setDonsCartItems] = useState({}); 
+  const [javaCartItems, setJavaCartItems] = useState({});        // Bon Bon's cart
 
   const PrivateRoute = ({ element, ...rest }) => {
     const { user } = useContext(AuthContext);
@@ -40,6 +45,7 @@ const App = () => {
   const [menuItems, setMenuItems] = useState({});
   const [bonsMenuItems, setBonsMenuItems] = useState({});
   const [donsMenuItems, setDonsMenuItems] = useState({});
+  const [javaMenuItems, setJavaMenuItems] = useState({});
 
    useEffect(() => {
     const fetchMenuItems = async () => {
@@ -53,7 +59,7 @@ const App = () => {
     };
     const fetchBonsMenuItems = async () => {
       try {
-        const response = await axios.get('http://localhost:5001/api/bonsMenuItems');
+        const response = await axios.get('http://localhost:5001/api/bons/menuItems');
         setBonsMenuItems(response.data);
       } catch (error) {
         console.error("Error fetching Bon Bon's menu items:", error);
@@ -62,7 +68,16 @@ const App = () => {
 
     const fetchDonsMenuItems = async () => {
       try {
-        const response = await axios.get('http://localhost:5001/api/donsMenuItems');
+        const response = await axios.get('http://localhost:5001/api/dons/menuItems');
+        setDonsMenuItems(response.data);
+      } catch (error) {
+        console.error("Error fetching Don's menu items:", error);
+      }
+    };
+
+    const fetchJavaMenuItems = async () => {
+      try {
+        const response = await axios.get('http://localhost:5001/api/java/menuItems');
         setDonsMenuItems(response.data);
       } catch (error) {
         console.error("Error fetching Don's menu items:", error);
@@ -73,6 +88,7 @@ const App = () => {
     fetchMenuItems();
     fetchBonsMenuItems();
     fetchDonsMenuItems();
+    fetchJavaMenuItems();
 }, []);
 
 
@@ -90,6 +106,7 @@ const App = () => {
           <Route path="/dashboard" element={<Dashboard />} />
           <Route path="/bonsdashboard" element={<BonsDashboard />} />
           <Route path="/donsdashboard" element={<DonsDashboard />} />
+          <Route path="/javadashboard" element={<JavaDashboard />} />
           <Route path="/checkout" element={<Checkout />} /> 
           <Route path="/orderConfirmation" element={<OrderConfirmation />} />
           <Route path='/trackOrder' element={<TrackOrder />} /> 
@@ -131,6 +148,17 @@ const App = () => {
                 <RestaurantsHeader cartItems={donsCartItems} setCartItems={setDonsCartItems} menuItems={donsMenuItems} restaurant={RestaurantsData["Don's at Walb"]} />
                 <DonsNavBar onCategorySelect={setDonsSelectedCategory} activeCategory={donsSelectedCategory} />
                 <DonsMenu category={donsSelectedCategory} cartItems={donsCartItems} setCartItems={setDonsCartItems} />
+              </>
+            } 
+          />
+          <Route 
+            path="/javamenu" 
+            element={
+              <> 
+                {/* <Header cartItems={bonsCartItems} setCartItems={setBonsCartItems} menuItems={bonsMenuItems} /> */}
+                <RestaurantsHeader cartItems={javaCartItems} setCartItems={setJavaCartItems} menuItems={javaMenuItems} restaurant={RestaurantsData["Java Spot"]} />
+                <JavaNavBar onCategorySelect={setJavaSelectedCategory} activeCategory={javaSelectedCategory} />
+                <JavaMenu category={javaSelectedCategory} cartItems={javaCartItems} setCartItems={setJavaCartItems} />
               </>
             } 
           />
